Refactoring LikelihoodSearchTable
