+++ traffic_stops/static/js/app/states/md/LikelihoodOfSearch.js
@@ -44,45 +44,13 @@ var LikelihoodOfSearch = C.LikelihoodOfSearchBase.extend({
   triggerRaceToggle: () => null
 });
 
-var LikelihoodSearchTable = TableBase.extend({
-  get_tabular_data: function () {
-    var header, row, rows = [];
+var LikelihoodSearchTable = C.LikelihoodSearchTableBase.extend({
+  types: [Stops.ethnicities],
 
-    // create header
-    header = ["Year", "Stop-reason"];
-    header.push.apply(header, Stops.ethnicities);
-    rows.push(header);
+  defaults: Stops,
 
-    var stop, search, stop_purp, search_purp, v1, v2,
-        purposes = Stops.purpose_order.keys(),
-        stops = this.data.raw.stops,
-        searches = this.data.raw.searches,
-        get_row = (stops, searches, term) => {
-          var stop = (stops !== undefined) ? stops[term] : 0,
-              search = (searches !== undefined) ? searches[term] : 0;
-          return `${search}/${stop}`;
-        };
-
-    // create data rows
-    this.data.years.forEach((yr) => {
-        stop = stops.filter((d) => d.year == yr);
-        search = searches.filter((d) => d.year == yr);
-        purposes.forEach((purp) => {
-          row = [yr, purp];
-          stop_purp = (stop.length>0) ? stop.filter((d) => d.purpose == purp): undefined;
-          search_purp = (search.length>0) ? search.filter((d) => d.purpose == purp) : undefined;
-          stop_purp = (stop_purp && stop_purp.length === 1) ? stop_purp[0] : undefined;
-          search_purp = (search_purp && search_purp.length === 1) ? search_purp[0] : undefined;
-
-          Stops.ethnicities.forEach((e) => {
-            row.push(get_row(stop_purp, search_purp, e));
-          });
-
-          rows.push(row);
-        });
-    });
-
-    return rows;
+  _get_header_rows: function () {
+    return Stops.ethnicities
   }
 });
