+++ traffic_stops/static/js/app/states/nc/LikelihoodOfSearch.js
@@ -47,49 +47,13 @@ var LikelihoodOfSearch = C.LikelihoodOfSearchBase.extend({
   }
 });
 
-var LikelihoodSearchTable = TableBase.extend({
-  get_tabular_data: function(){
-    var header, row, rows = [];
+var LikelihoodSearchTable = C.LikelihoodSearchTableBase.extend({
+  types: [Stops.races, Stops.ethnicities],
 
-    // create header
-    header = ["Year", "Stop-reason"];
-    header.push.apply(header, Stops.pprint.values());
-    rows.push(header);
+  defaults: Stops,
 
-    var stop, search, stop_purp, search_purp, v1, v2,
-        purposes = Stops.purpose_order.keys(),
-        stops = this.data.raw.stops,
-        searches = this.data.raw.searches,
-        get_row = function(stops, searches, term){
-          var stop = (stops !== undefined) ? stops[term] : 0,
-              search = (searches !== undefined) ? searches[term] : 0;
-          return `${search}/${stop}`;
-        };
-
-    // create data rows
-    this.data.years.forEach(function(yr){
-        stop = stops.filter(function(d){return d.year == yr;});
-        search = searches.filter(function(d){return d.year == yr;});
-        purposes.forEach(function(purp){
-          row = [yr, purp];
-          stop_purp = (stop.length>0) ? stop.filter(function(d){return d.purpose == purp;}): undefined;
-          search_purp = (search.length>0) ? search.filter(function(d){return d.purpose == purp;}) : undefined;
-          stop_purp = (stop_purp && stop_purp.length === 1) ? stop_purp[0] : undefined;
-          search_purp = (search_purp && search_purp.length === 1) ? search_purp[0] : undefined;
-
-          Stops.races.forEach(function(r){
-            row.push(get_row(stop_purp, search_purp, r));
-          });
-
-          Stops.ethnicities.forEach(function(e){
-            row.push(get_row(stop_purp, search_purp, e));
-          });
-
-          rows.push(row);
-        });
-    });
-
-    return rows;
+  _get_header_rows: function () {
+    return Stops.pprint.values();
   }
 });
